requirements.txt
# ─────────────────────────────────────────────────────────────
# Chronic Disease Chatbot – Dependency Manifest
# Architecture: LangGraph + Gemini + Tavily + ChromaDB + SQLite
# ─────────────────────────────────────────────────────────────

# ── Core Orchestration ────────────────────────────────────────
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0      # Gemini Pro / Flash

# ── LLM Provider ─────────────────────────────────────────────
google-generativeai>=0.7.0         # Underlying Gemini SDK

# ── Web Search (Knowledge Agent) ─────────────────────────────
tavily-python>=0.3.0               # Tavily Search API – LLM-optimised results

# ── Vector Memory (Memory Agent) ─────────────────────────────
chromadb>=0.5.0                    # Local vector store (includes ONNX-based default embeddings)

# ── Structured DB (Memory Agent) ─────────────────────────────
# sqlite3 is part of Python stdlib – no install needed

# ── Action Tools ─────────────────────────────────────────────
google-api-python-client>=2.130.0  # Google Calendar API
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
# Email: uses Python stdlib smtplib + Gmail App Password — no package needed

# ── Web / API Framework ───────────────────────────────────────
fastapi>=0.111.0                   # REST API layer for the chatbot
uvicorn[standard]>=0.29.0          # ASGI server
python-multipart>=0.0.9            # Form data support

# ── Environment & Config ──────────────────────────────────────
python-dotenv>=1.0.0               # Load .env secrets
pydantic>=2.7.0                    # Data validation / settings

# ── Utilities ─────────────────────────────────────────────────
rich>=13.7.0                       # Pretty console output / logging
tenacity>=8.3.0                    # Retry logic for API calls
httpx>=0.27.0                      # Async HTTP client

# ── Dev / Testing ─────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
